[SPARK-35832][CORE][K8S][TESTS] Add LocalRootDirsTest trait
